Move some methods to `diagnostics.rs` away from `parser.rs`
Move a bunch of error recovery methods to `diagnostics.rs` away from
`parser.rs`.
